feat: resolve page flux and move report page to another file
