+++ src/app-rust/Cargo.toml
@@ -9,6 +9,7 @@ printpdf = "0.3.1"
 rfd = "0.15.3"
 pulldown-cmark = "0.9"
 tempfile = "3.20.0"
+dioxus-router = "0.6.3"
 
 [features]
 default = ["web"]
