--- src/app-rust/src/Report/report_page.rs
@@ -1,1 +1,0 @@
-use dioxus::prelude::*;
